+++ changelog.md
@@ -41,6 +41,7 @@
   incoming publishes.
 - Enable to define the MQTT protocol version to be used by vmq_bridge (4 or 3).
 - Fix vmq_bridge dynamic configuration handling.
+- Fix multiple minor issues in vmq_swc.
 
 ## VerneMQ 1.9.2